List Projects

+++ app/Config/Routes.php
@@ -16,8 +16,8 @@ $routes->group('api', ['namespace' => 'App\Controllers\Api'], static function ($
     $routes->get('logout', 'AuthController::logout', ['filter' => 'apiauth']);
 
     $routes->post('add-project', 'ProjectController::addProject', ['filter' => 'apiauth']);
-    $routes->post('list-projects', 'ProjectController::listProjects');
-    $routes->delete('delete-project/(:any)', 'ProjectController::deleteProject');
+    $routes->get('list-projects', 'ProjectController::listProjects', ['filter' => 'apiauth']);
+    $routes->delete('delete-project/(:any)', 'ProjectController::deleteProject', ['filter' => 'apiauth']);
 
     $routes->get('invalid', 'AuthController::invalidRequest');
 });
